refactor: use generation for enroll, merge enroll component
