+++ src/locales/ar/builder.json
@@ -497,9 +497,6 @@
   "MEMBERSHIP_REQUEST_ACCEPT_BUTTON": "يقبل",
   "MEMBERSHIP_REQUEST_REJECT_BUTTON": "يرفض",
   "MEMBERSHIP_REQUESTS_TABLE_EMPTY": "لا يوجد طلب معلق.",
-  "ENROLL_TITLE": "التسجيل لهذا العنصر",
-  "ENROLL_BUTTON": "تسجل",
-  "ENROLL_DESCRIPTION": "انقر على الزر أدناه للتسجيل والوصول إلى هذا العنصر.",
   "REQUEST_ACCESS_PENDING_TITLE": "طلب الوصول الخاص بك معلق",
   "REQUEST_ACCESS_PENDING_DESCRIPTION": "يحتاج المسؤول إلى الموافقة على طلبك لتتمكن من الوصول إلى هذا العنصر.",
   "REQUEST_ACCESS_TITLE": "طلب الوصول إلى هذا العنصر",
