+++ src/locales/ar/common.json
@@ -58,5 +58,10 @@
     "MY_GRAASP": "العناصر الخاصّة بي",
     "BOOKMARKED_ITEMS": "العناصر المفضلة",
     "TRASH": "القمامة"
+  },
+  "ENROLL": {
+    "TITLE": "التسجيل لهذا العنصر",
+    "BUTTON": "تسجل",
+    "DESCRIPTION": "انقر على الزر أدناه للتسجيل والوصول إلى هذا العنصر."
   }
 }
