+++ src/locales/ar/player.json
@@ -38,9 +38,6 @@
   "REQUEST_ACCESS_TITLE": "طلب الوصول إلى هذا العنصر",
   "REQUEST_ACCESS_BUTTON": "طلب الوصول",
   "REQUEST_ACCESS_SENT_BUTTON": "تم إرسال الطلب",
-  "ENROLL_TITLE": "التسجيل لهذا العنصر",
-  "ENROLL_BUTTON": "تسجل",
-  "ENROLL_DESCRIPTION": "انقر على الزر أدناه للتسجيل والوصول إلى هذا العنصر.",
   "ITEM_LOGIN_HELPER_SIGN_OUT": "لقد قمت بتسجيل الدخول باسم {{email}}",
   "HOME_EMPTY": "ليس لديك أي عناصر.",
   "AUTO_LOGIN_NO_ITEM_LOGIN_ERROR": "هذا العنصر لا يدعم تسجيل الدخول التلقائي",
