+++ src/locales/de/builder.json
@@ -509,9 +509,6 @@
   "MEMBERSHIP_REQUEST_ACCEPT_BUTTON": "Akzeptieren",
   "MEMBERSHIP_REQUEST_REJECT_BUTTON": "Ablehnen",
   "MEMBERSHIP_REQUESTS_TABLE_EMPTY": "Keine ausstehende Anfrage.",
-  "ENROLL_TITLE": "Für diesen Artikel anmelden",
-  "ENROLL_BUTTON": "Einschreiben",
-  "ENROLL_DESCRIPTION": "Klicken Sie auf die Schaltfläche unten, um sich anzumelden und auf diesen Artikel zuzugreifen.",
   "REQUEST_ACCESS_PENDING_TITLE": "Ihre Zugriffsanfrage steht noch aus",
   "REQUEST_ACCESS_PENDING_DESCRIPTION": "Ein Administrator muss Ihre Anfrage genehmigen, damit Sie auf dieses Element zugreifen können.",
   "REQUEST_ACCESS_TITLE": "Zugriff auf diesen Artikel anfordern",
