+++ src/locales/de/player.json
@@ -38,9 +38,6 @@
   "REQUEST_ACCESS_TITLE": "Zugriff auf diesen Artikel anfordern",
   "REQUEST_ACCESS_BUTTON": "Zugriff beantragen",
   "REQUEST_ACCESS_SENT_BUTTON": "Anfrage gesendet",
-  "ENROLL_TITLE": "Für diesen Artikel anmelden",
-  "ENROLL_BUTTON": "Einschreiben",
-  "ENROLL_DESCRIPTION": "Klicken Sie auf die Schaltfläche unten, um sich anzumelden und auf diesen Artikel zuzugreifen.",
   "ITEM_LOGIN_HELPER_SIGN_OUT": "Sie sind angemeldet als {{email}}",
   "HOME_EMPTY": "Sie haben keine Elemente.",
   "AUTO_LOGIN_NO_ITEM_LOGIN_ERROR": "Dieser Artikel unterstützt keine automatische Anmeldung",
