+++ src/locales/en/builder.json
@@ -507,9 +507,6 @@
   "MEMBERSHIP_REQUEST_ACCEPT_BUTTON": "Accept",
   "MEMBERSHIP_REQUEST_REJECT_BUTTON": "Reject",
   "MEMBERSHIP_REQUESTS_TABLE_EMPTY": "No pending request.",
-  "ENROLL_TITLE": "Enroll to this item",
-  "ENROLL_BUTTON": "Enroll",
-  "ENROLL_DESCRIPTION": "Click on the button below to enroll and access this item.",
   "REQUEST_ACCESS_PENDING_TITLE": "Your access request is pending",
   "REQUEST_ACCESS_PENDING_DESCRIPTION": "An admin needs to approve your request for you to access this item.",
   "REQUEST_ACCESS_TITLE": "Request access to this item",
