+++ src/locales/en/common.json
@@ -73,5 +73,10 @@
     "REMOVE_BOOKMARK": "Remove bookmark",
     "ADD_LOCATION": "Add at this location"
   },
-  "FOLDER_HEADER_CREATED_BY": "Created by {{name}}"
+  "FOLDER_HEADER_CREATED_BY": "Created by {{name}}",
+  "ENROLL": {
+    "TITLE": "Enroll to this item",
+    "BUTTON": "Enroll",
+    "DESCRIPTION": "Click on the button below to enroll and access this item."
+  }
 }
