+++ src/locales/en/messages.json
@@ -153,6 +153,7 @@
   "REQUEST_MEMBERSHIP": "Membership requests has been sent successfully!",
   "DELETE_MEMBERSHIP_REQUEST": "Membership request has been successfully deleted.",
   "ENROLL": "You successfully enrolled to the item.",
+  "ENROLL_UNEXPECTED_ERROR": "An error occurred when enrolling.",
   "CANNOT_REGISTER_ON_FROZEN_ITEM_LOGIN_SCHEMA": "Cannot register on the targeted item because the subscription has been frozen by its owner",
   "CANNOT_ENROLL_FROZEN_ITEM_LOGIN_SCHEMA": "Cannot enroll in the targeted item because the subscription has been frozen by its owner",
   "NOT_MEMBER_OR_GUEST": "Your account is not registered as member or pseudonimized user",
