+++ src/locales/en/player.json
@@ -38,9 +38,6 @@
   "REQUEST_ACCESS_TITLE": "Request access to this item",
   "REQUEST_ACCESS_BUTTON": "Request access",
   "REQUEST_ACCESS_SENT_BUTTON": "Request sent",
-  "ENROLL_TITLE": "Enroll to this item",
-  "ENROLL_BUTTON": "Enroll",
-  "ENROLL_DESCRIPTION": "Click on the button below to enroll and access this item.",
   "ITEM_LOGIN_HELPER_SIGN_OUT": "You're logged in as {{email}}",
   "HOME_EMPTY": "You have no elements.",
   "AUTO_LOGIN_NO_ITEM_LOGIN_ERROR": "This item does not support auto-login",
