+++ src/locales/es/builder.json
@@ -509,9 +509,6 @@
   "MEMBERSHIP_REQUEST_ACCEPT_BUTTON": "Aceptar",
   "MEMBERSHIP_REQUEST_REJECT_BUTTON": "Rechazar",
   "MEMBERSHIP_REQUESTS_TABLE_EMPTY": "No hay ninguna solicitud pendiente.",
-  "ENROLL_TITLE": "Inscríbete a este artículo",
-  "ENROLL_BUTTON": "Inscribirse",
-  "ENROLL_DESCRIPTION": "Haga clic en el botón a continuación para inscribirse y acceder a este artículo.",
   "REQUEST_ACCESS_PENDING_TITLE": "Su solicitud de acceso está pendiente",
   "REQUEST_ACCESS_PENDING_DESCRIPTION": "Un administrador debe aprobar su solicitud para que usted pueda acceder a este elemento.",
   "REQUEST_ACCESS_TITLE": "Solicitar acceso a este artículo",
