+++ src/locales/es/common.json
@@ -67,5 +67,10 @@
     "LOG_IN": "Acceso",
     "LOG_OUT": "Finalizar la sesión",
     "NEED_HELP": "¿Necesitar ayuda?"
+  },
+  "ENROLL": {
+    "TITLE": "Inscríbete a este artículo",
+    "BUTTON": "Inscribirse",
+    "DESCRIPTION": "Haga clic en el botón a continuación para inscribirse y acceder a este artículo."
   }
 }
