+++ src/locales/es/player.json
@@ -38,9 +38,6 @@
   "REQUEST_ACCESS_TITLE": "Solicitar acceso a este artículo",
   "REQUEST_ACCESS_BUTTON": "Solicitar acceso",
   "REQUEST_ACCESS_SENT_BUTTON": "Solicitud enviada",
-  "ENROLL_TITLE": "Inscríbete a este artículo",
-  "ENROLL_BUTTON": "Inscribirse",
-  "ENROLL_DESCRIPTION": "Haga clic en el botón a continuación para inscribirse y acceder a este artículo.",
   "ITEM_LOGIN_HELPER_SIGN_OUT": "Has iniciado sesión como {{email}}",
   "HOME_EMPTY": "No tienes elementos",
   "AUTO_LOGIN_NO_ITEM_LOGIN_ERROR": "Este artículo no admite el inicio de sesión automático",
