+++ src/locales/fr/builder.json
@@ -507,9 +507,6 @@
   "MEMBERSHIP_REQUEST_ACCEPT_BUTTON": "Accepter",
   "MEMBERSHIP_REQUEST_REJECT_BUTTON": "Refuser",
   "MEMBERSHIP_REQUESTS_TABLE_EMPTY": "Aucune requête en attente.",
-  "ENROLL_TITLE": "Rejoindre cet élément",
-  "ENROLL_BUTTON": "Rejoindre",
-  "ENROLL_DESCRIPTION": "Cliquez sur le bouton ci-dessous pour rejoindre et accéder à cet élément.",
   "REQUEST_ACCESS_PENDING_TITLE": "Votre demande d'accès est en cours",
   "REQUEST_ACCESS_PENDING_DESCRIPTION": "Un administrateur doit approuver votre requête pour que vous puissiez y accéder.",
   "REQUEST_ACCESS_TITLE": "Demander l'accès à cet élément",
