+++ src/locales/fr/player.json
@@ -38,9 +38,6 @@
   "REQUEST_ACCESS_TITLE": "Demander l'accès à cette capsule",
   "REQUEST_ACCESS_BUTTON": "Demande d'accès",
   "REQUEST_ACCESS_SENT_BUTTON": "Demande envoyée",
-  "ENROLL_TITLE": "Inscrivez-vous à cette capsule",
-  "ENROLL_BUTTON": "S'inscrire",
-  "ENROLL_DESCRIPTION": "Cliquez sur le bouton ci-dessous pour vous inscrire et accéder à cette capsule.",
   "ITEM_LOGIN_HELPER_SIGN_OUT": "Vous êtes connecté en tant que {{email}}",
   "HOME_EMPTY": "Vous n'avez aucun élément.",
   "AUTO_LOGIN_NO_ITEM_LOGIN_ERROR": "Cet article ne prend pas en charge la connexion automatique",
