+++ src/locales/it/builder.json
@@ -499,9 +499,6 @@
   "MEMBERSHIP_REQUEST_ACCEPT_BUTTON": "Accettare",
   "MEMBERSHIP_REQUEST_REJECT_BUTTON": "Rifiutare",
   "MEMBERSHIP_REQUESTS_TABLE_EMPTY": "Nessuna richiesta in sospeso.",
-  "ENROLL_TITLE": "Iscriviti a questo articolo",
-  "ENROLL_BUTTON": "Iscriversi",
-  "ENROLL_DESCRIPTION": "Clicca sul pulsante qui sotto per registrarti e accedere a questo articolo.",
   "REQUEST_ACCESS_PENDING_TITLE": "La tua richiesta di accesso è in sospeso",
   "REQUEST_ACCESS_PENDING_DESCRIPTION": "Per poter accedere a questo elemento, la tua richiesta deve essere approvata da un amministratore.",
   "REQUEST_ACCESS_TITLE": "Richiedi l'accesso a questo elemento",
