+++ src/locales/it/player.json
@@ -38,9 +38,6 @@
   "REQUEST_ACCESS_TITLE": "Richiedi l'accesso a questo elemento",
   "REQUEST_ACCESS_BUTTON": "Richiedi accesso",
   "REQUEST_ACCESS_SENT_BUTTON": "Richiesta inviata",
-  "ENROLL_TITLE": "Iscriviti a questo articolo",
-  "ENROLL_BUTTON": "Iscriversi",
-  "ENROLL_DESCRIPTION": "Clicca sul pulsante qui sotto per registrarti e accedere a questo articolo.",
   "ITEM_LOGIN_HELPER_SIGN_OUT": "Hai effettuato l'accesso come {{email}}",
   "HOME_EMPTY": "Non hai elementi.",
   "AUTO_LOGIN_NO_ITEM_LOGIN_ERROR": "Questo elemento non supporta l'accesso automatico",
