--- src/query/item/itemLogin/routes.ts
@@ -1,6 +1,0 @@
-import { DiscriminatedItem } from '@graasp/sdk';
-
-import { ITEMS_ROUTE } from '../routes.js';
-
-export const buildEnroll = (itemId: DiscriminatedItem['id']) =>
-  `${ITEMS_ROUTE}/${itemId}/enroll`;
