--- src/query/item/itemLogin/routines.ts
@@ -1,3 +1,0 @@
-import createRoutine from '../../routines/utils.js';
-
-export const enrollRoutine = createRoutine('ENROLL');
