+++ src/query/mutations/itemLogin.ts
@@ -3,7 +3,6 @@ import { ItemLoginSchemaStatus, ItemLoginSchemaType, UUID } from '@graasp/sdk';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 
 import * as Api from '../api/itemLogin.js';
-import { useEnroll } from '../item/itemLogin/mutations.js';
 import { itemKeys } from '../keys.js';
 import {
   deleteItemLoginSchemaRoutine,
@@ -99,7 +98,6 @@ export default (queryConfig: QueryClientConfig) => {
   };
 
   return {
-    useEnroll: useEnroll(queryConfig),
     usePostItemLogin,
     usePutItemLoginSchema,
     useDeleteItemLoginSchema,
